+++ app/http/controllers/projects.go
@@ -52,14 +52,23 @@ func GetProjects(request contracts.HttpRequest, guard contracts.Guard) any {
 	}
 }
 
-func GetProject(request contracts.HttpRequest) any {
+func GetProject(request contracts.HttpRequest, guard contracts.Guard) any {
+	project := models.Projects().FindOrFail(request.GetString("id"))
+	if !usecase.HasProjectPermission(project, utils.ToInt(guard.GetId(), 0)) {
+		return contracts.Fields{"msg": "没有该项目的权限"}
+	}
 	return contracts.Fields{
-		"data": usecase.GetProjectDetail(request.GetString("id")),
+		"data": usecase.GetProjectDetail(project),
 	}
 }
 
-func DeleteProject(request contracts.HttpRequest) any {
+func DeleteProject(request contracts.HttpRequest, guard contracts.Guard) any {
 	project := models.Projects().FindOrFail(request.Get("id"))
+
+	if !usecase.HasProjectPermission(project, utils.ToInt(guard.GetId(), 0)) {
+		return contracts.Fields{"msg": "没有该项目的权限"}
+	}
+
 	err := usecase.DeleteProject(project)
 	if err != nil {
 		return contracts.Fields{"msg": "删除失败：" + err.Error()}
@@ -70,6 +79,10 @@ func DeleteProject(request contracts.HttpRequest) any {
 func CreateProject(request requests.ProjectRequest, guard contracts.Guard) any {
 	validation.VerifyForm(request)
 	fields := request.Fields()
+
+	if groupId := utils.ToInt(fields["group_id"], 0); groupId > 0 && !usecase.HasGroupPermission(models.Groups().FindOrFail(groupId), utils.ToInt(guard.GetId(), 0)) {
+		return contracts.Fields{"msg": "没有该分组的权限"}
+	}
 
 	project, err := usecase.CreateProject(guard.GetId(), fields)
 
@@ -117,9 +130,13 @@ func CopyProject(request contracts.HttpRequest, guard contracts.Guard) any {
 	}
 }
 
-func UpdateProject(request requests.ProjectRequest) any {
+func UpdateProject(request requests.ProjectRequest, guard contracts.Guard) any {
 	validation.VerifyForm(request)
 	fields := request.Fields()
+	targetProject := models.Projects().FindOrFail(request.GetString("id"))
+	if !usecase.HasProjectPermission(targetProject, utils.ToInt(guard.GetId(), 0)) {
+		return contracts.Fields{"msg": "没有该项目的权限"}
+	}
 	project, err := usecase.UpdateProject(request.GetInt("id"), fields)
 	if err != nil {
 		return contracts.Fields{"msg": err.Error()}
